// src/globals/enums.ts
export enum BrickId {
  About = "about",
  Content = "content",
  Footer = "footer",
  Gallery = "gallery",
  Header = "header",
  Hero = "hero",
  Quote = "quote",
  SocialMediaFeed = "socialMediaFeed"
}

export enum BuilderMode {
  Add = "add",
  Edit = "edit",
  Image = "image",
  View ="view"
}

export enum PreviewMode {
  Desktop = "desktop",
  Mobile = "mobile"
}

export enum ViewBricksMenuMode {
  BrickCopy = "Brick Copy",
  BrickOptions = "Brick Options",
  ImageOptions = "Image Options"
}

export enum EditBrickMode {
  Copy = "Copy",
  Image = "Image",
  None = "None"
}

export enum FontFamily {
  Arvo = "Arvo",
  Anton_Thick = "Anton Thick",
  Baskerville = "Baskerville",
  Inter = "Inter",
  Lora = "Lora",
  Merriweather = "Merriweather",
  Montserrat = "Montserrat",
  Open_Sans = "Open Sans",
  Poppins = "Poppins",
  PT_Serif = "PT Serif",
  Roboto = "Roboto",
  Sans_Serif = "Sans Serif",
  Serif = "Serif"
}

export enum FontWeight {
  Bold = "bold",
  Normal = "normal"
}

export enum FontStyle {
  Italic = "italic",
  Normal = "normal"
}

export enum TextDecoration {
  Underline = "underline",
  None = "none"
}

export enum TextTransform {
  Uppercase = "uppercase",
  None = "none"
}

export enum TextAlign {
  Left = "left",
  Center = "center",
  Right = "right",
  SpaceBetween = "space-between"
}

export enum Color {
  Black = "#000000",
  Blue = "#3392CE",
  BlueHover = "#43A2DE",
  BlueGray = "#E9F1FE",
  DarkBlue = "#2E6188",
  DarkerBlue = "#285F88",
  DarkGreen = "#003D3E",
  Gray1 = "#F9FAFC",
  Gray2 = "#F7F7F7",
  Gray3 = "#E6E5E5",
  Gray4 = "#DCDCDE",
  Gray5 = "#BBC1CB",
  Gray6 = "#A8AAAD",
  Gray7 = "#9F9F9F",
  Gray8 = "#6F6F6F",
  Gray9 = "#333333",
  Gray10 = "#1F1F1F",
  Green = "#67AE5B",
  LightBlue = "#4F90C9",
  LightOrange = "#FCF1EB",
  Orange = "#FF7F00",
  Red = "#EF1552",
  RedOrange = "#EA6E32",
  Turquoise = "#54BAD2",
  White = "#FFFFFF",
  Yellow = "#FCEB61",
  YellowGreen = "#D0DB59",
}

export enum LinksType {
  List = "list",
  Menu = "menu"
}

export enum GalleryType {
  Carousel = "carousel",
  Grid = "grid",
  Minimized = "minimized"
}

export enum SocialMediaType {
  Instagram = "instagram",
  Twitter = "twitter",
  TwitterInstagram = "twitter_instagram"
}